# tacocloud/tacocloud-api/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Taco Cloud Reactive API
  description: Reactive Taco Cloud REST API supporting non-blocking order placement, pricing, inventory reservation, idempotency, and transactional outbox.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/v1/users/me/orders:
    get:
      summary: Get paginated order history of current user
      description: Returns authenticated user's private order history.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Paginated orders
          content:
            application/json:
              schema:
                type: object
...